rjson refactoring, porting to llvm 10.0

+++ cc/antigenic-maps.cc
@@ -29,7 +29,7 @@ void acmacs::tal::v3::AntigenicMaps::columns_rows()
         //     rows_ = num_sections;
         // else
         rows_ = 3;
-        width_to_height_ratio() = columns_ / 3.0;
+        width_to_height_ratio() = static_cast<double>(columns_) / 3.0;
     }
     else {
         columns_ = 0;
@@ -46,11 +46,11 @@ void acmacs::tal::v3::AntigenicMaps::draw(acmacs::surface::Surface& surface) con
     if (columns_ && rows_) {
         const auto& viewport = surface.viewport();
         const auto gap = surface.convert(Pixels{parameters().gap_between_maps}).value();
-        const auto map_size = viewport.size.height / rows_ - gap * (rows_ - 1) / rows_;
+        const auto map_size = viewport.size.height / static_cast<double>(rows_) - gap * static_cast<double>(rows_ - 1) / static_cast<double>(rows_);
         const auto* hz_sections = tal().draw().layout().find<HzSections>();
         for (auto [section_no, section] : acmacs::enumerate(hz_sections->sections())) {
-            const auto left = (section_no % columns_) * (map_size + gap);
-            const auto top = (section_no / columns_) * (map_size + gap);
+            const auto left = static_cast<double>(section_no % columns_) * (map_size + gap);
+            const auto top = static_cast<double>(section_no / columns_) * (map_size + gap);
             draw_map(surface.subsurface({left, top}, Scaled{map_size}, Size{1.0, 1.0}, false), section);
         }
     }
